apps/snowflake-product/requirements.txt
appdirs==1.4.4
asn1crypto==1.3.0
attrs==19.3.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
black==19.10b0
blis==0.4.1
boto3==1.11.17
botocore==1.14.17
Brotli==1.0.7
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
cymem==2.0.3
dash==1.12.0
dash-bootstrap-components==0.9.2
dash-core-components==1.10.0
dash-dangerously-set-inner-html==0.0.2
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
docutils==0.15.2
Flask==1.1.2
Flask-Caching==1.8.0
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
ijson==2.6.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.14.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
murmurhash==1.0.2
nltk==3.5
numpy==1.18.3
oscrypto==1.2.0
pandas==1.0.3
pathspec==0.8.0
plac==1.1.3
plotly==4.7.1
preshed==3.0.2
pyarrow==0.16.0
pycparser==2.20
pycryptodomex==3.9.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
regex==2020.4.4
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
six==1.14.0
snowflake-connector-python==2.2.5
snowflake-sqlalchemy==1.2.3
spacy==2.2.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
SQLAlchemy==1.3.16
srsly==1.0.2
textblob==0.15.3
thinc==7.4.0
toml==0.10.1
tqdm==4.45.0
typed-ast==1.4.1
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
zipp==3.1.0
